pyproject.toml
[tool.poetry]
name = "kindel"
version = "0.1.0"
description = ""
authors = ["A. F. Neto <51216181+hzeri@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
xgboost = ">=2.1.1,<3"
redun = ">=0.22.0,<0.23"
pandas = "*"
rdkit = "*"
tqdm = ">=4.66.5,<5"
scikit-learn = "*"
pyarrow = "*"
numpy = "<2"
wandb = "*"
omegaconf = ">=2.3.0,<3"
python-dotenv = ">=1.0.1,<2"
torch = {version = "^2.5.1+cu124", source = "torchgpu"}
torch-geometric = ">=2.5.2,<3"
lightning = ">=2.4.0,<3"
matplotlib = "^3.9.2"


[[tool.poetry.source]]
name = "torchgpu"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
